Refactor game_list_frame
