+++ rpcs3/rpcs3.vcxproj.filters
@@ -1254,6 +1254,24 @@
     <ClCompile Include="Input\sdl_camera_video_sink.cpp">
       <Filter>Io\camera</Filter>
     </ClCompile>
+    <ClCompile Include="rpcs3qt\game_list_context_menu.cpp">
+      <Filter>Gui\game list</Filter>
+    </ClCompile>
+    <ClCompile Include="rpcs3qt\game_list_actions.cpp">
+      <Filter>Gui\game list</Filter>
+    </ClCompile>
+    <ClCompile Include="QTGeneratedFiles\Debug\moc_game_list_context_menu.cpp">
+      <Filter>Generated Files\Debug</Filter>
+    </ClCompile>
+    <ClCompile Include="QTGeneratedFiles\Debug\moc_game_list_actions.cpp">
+      <Filter>Generated Files\Debug</Filter>
+    </ClCompile>
+    <ClCompile Include="QTGeneratedFiles\Release\moc_game_list_actions.cpp">
+      <Filter>Generated Files\Release</Filter>
+    </ClCompile>
+    <ClCompile Include="QTGeneratedFiles\Release\moc_game_list_context_menu.cpp">
+      <Filter>Generated Files\Release</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Input\ds4_pad_handler.h">
@@ -1836,6 +1854,12 @@
     </CustomBuild>
     <CustomBuild Include="rpcs3qt\debugger_add_bp_window.h">
       <Filter>Gui\debugger</Filter>
+    </CustomBuild>
+    <CustomBuild Include="rpcs3qt\game_list_context_menu.h">
+      <Filter>Gui\game list</Filter>
+    </CustomBuild>
+    <CustomBuild Include="rpcs3qt\game_list_actions.h">
+      <Filter>Gui\game list</Filter>
     </CustomBuild>
   </ItemGroup>
   <ItemGroup>
